+++ libgammu/phone/obex/obexfunc.h
@@ -10,9 +10,9 @@
 extern GSM_Reply_Function	OBEXGENReplyFunctions[];
 extern GSM_Error OBEXGEN_InitialiseVars(GSM_StateMachine *s);
 extern void OBEXGEN_FreeVars(GSM_StateMachine *s);
-extern GSM_Error OBEXGEN_GetFilePart	(GSM_StateMachine *s, GSM_File *File, int *Handle, int *Size);
-extern GSM_Error OBEXGEN_AddFilePart	(GSM_StateMachine *s, GSM_File *File, int *Pos, int *Handle);
-extern GSM_Error OBEXGEN_SendFilePart	(GSM_StateMachine *s, GSM_File *File, int *Pos, int *Handle);
+extern GSM_Error OBEXGEN_GetFilePart	(GSM_StateMachine *s, GSM_File *File, int *Handle, size_t *Size);
+extern GSM_Error OBEXGEN_AddFilePart	(GSM_StateMachine *s, GSM_File *File, size_t *Pos, int *Handle);
+extern GSM_Error OBEXGEN_SendFilePart	(GSM_StateMachine *s, GSM_File *File, size_t *Pos, int *Handle);
 extern GSM_Error OBEXGEN_GetNextFileFolder(GSM_StateMachine *s, GSM_File *File, gboolean start);
 extern GSM_Error OBEXGEN_Disconnect	(GSM_StateMachine *s);
 extern GSM_Error OBEXGEN_Connect	(GSM_StateMachine *s, OBEX_Service service);
